+++ test/run.sh
@@ -3,7 +3,6 @@
 # bail on error
 set -e
 
-# TODO(sigmund): replace with a real test runner
 DIR=$( cd $( dirname "${BASH_SOURCE[0]}" ) && pwd )
 
 # Note: dart_analyzer needs to be run from the root directory for proper path